django-account/requirements.txt
django
djangorestframework
python-decouple
django-cors-headers
psycopg2-binary
django-s3-storage